+++ backend/utils/errorflags.js
@@ -1,0 +1,7 @@
+export default userInput_ErrFlags = {
+    1: "First name cannot be empty or contain numbers",
+    2: "Last name cannot be empty or contain numbers",
+    3: "Invalid email",
+    4: "Email already registered. Login instead",
+  };
+
